+++ NEWS.md
@@ -1,3 +1,12 @@
-# jenga 1.0.1
+# jenga 1.0.0
 
 * Added a `NEWS.md` file to track changes to the package.
+
+# jenga 1.1.0
+
+* common sense filter
+* added stats to the prediction table (and removed pred_stats)
+* added new criterion for model selection: prediction score
+* added new metric measures for test error from greybox
+* added automatic differentiation module (and removed deriv from hyper-parameters)
+